Title: 괄호 쌍 만들어주기, Time: 255ms, Memory: 14MB, Status: Passed - CodeTree

+++ 250313/괄호 쌍 만들어주기/pair-parentheses.java
@@ -3,11 +3,11 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         String a = sc.next();
-        int[] arr = new int[a.length()];
+        long[] arr = new long[a.length()];
         boolean check = false;
         boolean count = false;
-        int result = 0;
-        int num = 0;
+        long result = 0;
+        long num = 0;
         for(int i=a.length()-1;i>=0;i--){
             char now = a.charAt(i);
             if(now==')'&&check){
